rename locking -> wallet
