+++ plugins/Wallet/js/constants/wallet.js
